# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "paths": {
        "/surveys": {
            "get": {
                "tags": [
                    "Surveys"
                ],
                "summary": "Get all surveys",
                "description": "Get all surveys that have status=1",
                "operationId": "getSurveys",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Survey"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Surveys"
                ],
                "summary": "Create a new survey",
                "description": "Create a new survey",
                "operationId": "createSurvey",
                "requestBody": {
                    "description": "Survey object to be created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SurveyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Survey"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/surveys/{survey}": {
            "get": {
                "tags": [
                    "Surveys"
                ],
                "summary": "Get a survey by ID",
                "description": "Get a survey by ID",
                "operationId": "getSurveyById",
                "parameters": [
                    {
                        "name": "survey",
                        "in": "path",
                        "description": "ID of the survey to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Survey"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "put": {
                "tags": [
                    "Surveys"
                ],
                "summary": "Update an existing survey",
                "description": "Update an existing survey",
                "operationId": "updateSurvey",
                "parameters": [
                    {
                        "name": "survey",
                        "in": "path",
                        "description": "ID of the survey to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Survey object to be updated",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SurveyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Survey"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Survey": {
                "title": "Survey",
                "required": [
                    "title",
                    "description",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            }
        }
    }
}